% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyff-package.r, R/tidyff.R
\docType{package}
\name{tidyff}
\alias{tidyff}
\alias{tidyff-package}
\alias{tidyff}
\title{tidyff.}
\usage{
tidyff(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{ignored}
}
\description{
tidyff.

Tidy ff
}
